longevity_project_structured_clean/requirements.txt
python-dotenv
requests
beautifulsoup4
pandas
aiohttp
urllib3
lxml
tenacity
sentence-transformers
jsonschema
PyYAML
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
banks==2.2.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
config==0.5.1
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distlib==0.4.0
distro==1.9.0
executing==2.2.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.4
griffe==1.11.1
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
identify==2.6.15
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.13.1
llama-index-cli==0.5.0
llama-index-core==0.14.5
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-nebius==0.4.0
llama-index-embeddings-openai==0.5.0
llama-index-indices-managed-llama-cloud==0.9.1
llama-index-instrumentation==0.4.0
llama-index-llms-huggingface==0.6.1
llama-index-llms-huggingface-api==0.6.1
llama-index-llms-nebius==0.2.0
llama-index-llms-ollama==0.8.0
llama-index-llms-openai==0.5.3
llama-index-llms-openai-like==0.5.0
llama-index-readers-file==0.5.1
llama-index-readers-llama-parse==0.5.0
llama-index-vector-stores-milvus==0.9.0
llama-index-workflows==2.8.1
llama-parse==0.6.54
lxml==6.0.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
milvus-lite==2.5.1
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
ollama==0.6.0
openai==1.99.9
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
pip==25.2
platformdirs==4.3.8
pre_commit==4.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.27.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pymilvus==2.6.0
pypdf==5.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
regex==2025.7.34
requests==2.32.4
ruff==0.12.0
sacremoses==0.1.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
triton==3.5.0
typing_extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
virtualenv==20.35.3
wcwidth==0.2.13
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1